changed test case location
